Update method calls to be asynchronous
The code changes involve updating the methods that retrieve the latest blockhash and the simulation transaction, to work asynchronously. By adding the await keyword, we ensure that execution is paused until the Promises they return are fulfilled. The methods are used in amm_math.rs and openbook.rs files in the AMM module of our library.